refactor(balance): split the account native balance map out of useTokenBalance
`useAccountListInfos` needs only the per-account balance map, but reaching it
through `useTokenBalance` also mounted the current account's Gno (5s) and Cosmos
(10s) balance queries. On the Accounts screen those are the only observers of
either query — wallet-main is unmounted — so opening the account list started
two pollers that nothing on that screen renders.

Move the map into its own hook. Mounting it is now the demand signal, so the
Recoil counter added earlier for the same purpose is no longer needed and goes
away along with its atom and the RecoilRoot the spec needed for it.

The query key, gating conditions and session/revoke balance-address policy are
unchanged; the map only ever fetched the native token, so it calls
`getGnotTokenBalance` directly instead of routing through the general
`fetchBalanceBy`.
